app/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
pandas
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
# Uncomment if using OTLP exporter:
# opentelemetry-exporter-otlp-proto-http
